test2/backend/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
requests==2.26.0
python-dotenv==0.19.0
paypayopa
